+++ src/utils.js
@@ -1,6 +1,13 @@
+'use strict'
+
+var utils = {
+    isFunction: function(arg) {
+        return typeof arg === 'function';
+    }
+};
 
 // Returns true if all digits in string s are valid hex numbers
-function IsValidHex(s) {
+utils.IsValidHex = function (s) {
     for (var i = 0; i < s.length; ++i) {
         var cc = s.charCodeAt(i);
         if ((48 <= cc && cc <= 57) ||
@@ -15,7 +22,7 @@ function IsValidHex(s) {
 }
 
 // Does the char code correspond to an alpha-numeric char.
-function isAlphaNumeric(cc) {
+utils.isAlphaNumeric = function isAlphaNumeric(cc) {
     // a - z
     if (97 <= cc && cc <= 122) return true;
     // A - Z
@@ -26,3 +33,19 @@ function isAlphaNumeric(cc) {
     return false;
 }
 
+/**
+ * 随机从列表返回指定长度的列表
+ * @param {array} array 待选列表
+ * @param {array} excludes 排除列表
+ * @param {number} len 返回的长度,默认 1
+ */
+utils.RandomChoice = function(array, excludes = [], len = 1) {
+    var tmp = [];
+    while (tmp.length < len) {
+        let v = array[Math.ceil(Math.random() * array.length - 1)];
+        excludes.indexOf(v) === -1 && tmp.indexOf(v) === -1 && tmp.push(v);
+    }
+    return tmp;
+}
+
+module.exports = utils;